% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about.R
\name{about}
\alias{about}
\title{Display information about the data source}
\usage{
about(display = c("html", "text"))
}
\arguments{
\item{display}{if \code{html} (the default) a formatted version of
the description is provided, otherwise a plaintext version
will be provided.}
}
\description{
Display information about the data source
}
