% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdccovidview-package.R
\docType{package}
\name{cdccovidview}
\alias{cdccovidview}
\alias{cdccovidview-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/cdccovidview}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/cdccovidview/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
