% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical-labs.R
\name{clinical_labs}
\alias{clinical_labs}
\title{Retrieve U.S. Clinical Laboratories Reporting SARS-CoV-2 Test Results to CDC}
\usage{
clinical_labs()
}
\value{
data frame
}
\description{
Retrieve U.S. Clinical Laboratories Reporting SARS-CoV-2 Test Results to CDC
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/04102020/reporting-cov2-results.html}
}
