% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_bom}
\alias{has_bom}
\title{Tests whether a raw httr response or character vector has a byte order mark (BOM)}
\usage{
has_bom(resp, encoding = "UTF-8")
}
\description{
Tests whether a raw httr response or character vector has a byte order mark (BOM)
}
