% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laboratory-confirmed-hospitalizations.R
\name{laboratory_confirmed_hospitalizations}
\alias{laboratory_confirmed_hospitalizations}
\title{Retrieve Laboratory-Confirmed COVID-19-Associated Hospitalizations}
\usage{
laboratory_confirmed_hospitalizations()
}
\value{
data frame
}
\description{
This function grabs all data for all networks, catchments, seasons, and ages.
In the future there will be ways of selecting just the desired target areas.
}
