% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.R
\name{mortality_surveillance_data}
\alias{mortality_surveillance_data}
\title{Retrieve NCHS Mortality Surveillance Data}
\usage{
mortality_surveillance_data()
}
\value{
data frame
}
\description{
Retrieve NCHS Mortality Surveillance Data
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/04102020/nchs-data.html}
}
