% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nssp-cd.R
\name{nssp_er_visits_national}
\alias{nssp_er_visits_national}
\title{Retrieve National Syndromic Surveillance Program (NSSP): Emergency Department Visits Percentage of Visits for COVID-19-Like Illness (CLI) or Influenza-like Illness (ILI)}
\usage{
nssp_er_visits_national()
}
\value{
data frame
}
\description{
Retrieve National Syndromic Surveillance Program (NSSP): Emergency Department Visits Percentage of Visits for COVID-19-Like Illness (CLI) or Influenza-like Illness (ILI)
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/04102020/nssp-regions.html}
}
