% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provisional-deaths.R
\name{provisional_death_counts}
\alias{provisional_death_counts}
\title{Retrieve Provisional Death Counts for Coronavirus Disease (COVID-19)}
\usage{
provisional_death_counts()
}
\value{
a list with 4 named elements: \code{by_week}, \code{by_age}, \code{by_state}, \code{by_sex}
}
\description{
Retrieve Provisional Death Counts for Coronavirus Disease (COVID-19)
}
\note{
Please see the indicated reference for all the caveats and precise meanings for each field. Also,
this function used the JSON API (\url{https://data.cdc.gov/resource/hc4f-j6nb.json})
}
\references{
\url{https://data.cdc.gov/api/views/hc4f-j6nb/rows.csv?accessType=DOWNLOAD&bom=true&format=true}
}
