% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-health-labs-national.R
\name{public_health_labs_national}
\alias{public_health_labs_national}
\title{Retrieve National Surveillance of U.S. State and Local Public Health Laboratories Reporting to CDC}
\usage{
public_health_labs_national()
}
\value{
data frame
}
\description{
Retrieve National Surveillance of U.S. State and Local Public Health Laboratories Reporting to CDC
}
\references{
\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/04102020/labs-regions.html}
}
