% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sans_bom}
\alias{sans_bom}
\title{Remove byte order mark (BOM) from \code{httr::response} object or character vector}
\usage{
sans_bom(resp)
}
\description{
Remove byte order mark (BOM) from \code{httr::response} object or character vector
}
