% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance-area-info.R
\name{surveillance_areas}
\alias{surveillance_areas}
\title{Show network & network catchments}
\usage{
surveillance_areas()
}
\value{
data frame
}
\description{
Show network & network catchments
}
