% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdccovidview-package.R
\docType{package}
\name{cdccovidview}
\alias{cdccovidview}
\alias{cdccovidview-package}
\title{Weekly Surveillance Summary of U.S. COVID-19 Activity}
\description{
The U.S. Centers for Disease Control provides weekly summary and
interpretation of key indicators that have been adapted to track the COVID-19
pandemic in the United States. Tools are provided to retrive data from both
COVIDView
(\url{https://www.cdc.gov/coronavirus/2019-ncov/covid-data/covidview/index.html})
and COVID-NET (\url{https://gis.cdc.gov/grasp/COVIDNet/COVID19_3.html}).
}
\details{
The Coronavirus Disease 2019 (COVID-19)-Associated Hospitalization
Surveillance Network (COVID-NET) conducts population-based surveillance for
laboratory-confirmed COVID-19-associated hospitalizations in children
(persons younger than 18 years) and adults. The current network covers nearly
100 counties in the 10 Emerging Infections Program (EIP) states (CA, CO, CT,
GA, MD, MN, NM, NY, OR, and TN) and four additional states through the
Influenza Hospitalization Surveillance Project (IA, MI, OH, and UT). The
network represents approximately 10\% of US population (~32 million people).

Cases are identified by reviewing hospital, laboratory, and admission
databases and infection control logs for patients hospitalized with a
documented positive SARS-CoV-2 test.

Data gathered are used to estimate age-specific hospitalization rates on a
weekly basis and describe characteristics of persons hospitalized with
COVID-19. Laboratory confirmation is dependent on clinician-ordered
SARS-CoV-2 testing. Therefore, the rates provided are likely to be
underestimated as COVID-19-associated hospitalizations can be missed due to
test availability and provider or facility testing practices.

COVID-NET hospitalization data are preliminary and subject to change as more
data become available. Please use the following citation when referencing
these data: “COVID-NET: COVID-19-Associated Hospitalization Surveillance
Network, Centers for Disease Control and Prevention. WEBSITE. Accessed on
DATE”.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/cdccovidview}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/cdccovidview/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
